SA-1328: Moved report builder files to separate folder
