+++ src/pages/reportBuilder/ReportBuilder.module.scss
@@ -1,0 +1,23 @@
+@import '~@sparkpost/matchbox/src/styles/config.scss';
+
+.ChartSection {
+  position: relative;
+  padding: spacing();
+  min-height: rem(300);
+
+  opacity: 1;
+  pointer-events: auto;
+  transition: 0.15s;
+}
+
+.pending {
+  opacity: 0.5;
+  pointer-events: none;
+}
+
+.Loading {
+  position: absolute;
+  top: 55%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
