+++ src/pages/reports/summary/SummaryPage.container.js
@@ -1,6 +1,6 @@
 import useHibanaToggle from 'src/hooks/useHibanaToggle';
 import SummaryPage from './SummaryPage';
-import ReportBuilder from './ReportBuilder';
+import ReportBuilder from '../../reportBuilder/ReportBuilder';
 
 export default function(props) {
   return useHibanaToggle(SummaryPage, ReportBuilder)(props);
